+++ tests/Unit/Tecan/TecanProtocolTest.php
@@ -2,6 +2,8 @@
 
 namespace Mll\LiquidHandlingRobotics\Tests\Unit\Tecan;
 
+use Carbon\Carbon;
+use Composer\InstalledVersions;
 use Illuminate\Support\Str;
 use Mll\LiquidHandlingRobotics\Tecan\BasicCommands\Aspirate;
 use Mll\LiquidHandlingRobotics\Tecan\BasicCommands\Dispense;
@@ -22,10 +24,33 @@ use PHPUnit\Framework\TestCase;
 
 final class TecanProtocolTest extends TestCase
 {
-    public function testProtocolCustomName(): void
+    public function setUp(): void
     {
-        $tecanProtocol = new TecanProtocol(TipMask::FOUR_TIPS(), 'testProtocol');
+        parent::setUp();
+        Carbon::setTestNow(Carbon::createStrict(2022, 10, 5, 13, 34, 32));
+    }
+
+    protected function tearDown(): void
+    {
+        parent::tearDown();
+        Carbon::setTestNow(Carbon::now());
+    }
+
+    public function testProtocolUserAndName(): void
+    {
+        $userName = 'username';
+        $protocolName = 'testProtocol';
+        $tecanProtocol = new TecanProtocol(TipMask::FOUR_TIPS(), $protocolName, $userName);
+
         self::assertSame('testProtocol.gwl', $tecanProtocol->fileName());
+        self::assertSame(
+            StringUtil::normalizeLineEndings(
+                $this->initComment() . 'C;User: username
+C;Protocol name: testProtocol
+'
+            ),
+            $tecanProtocol->buildProtocol()
+        );
     }
 
     public function testProtocolUuidName(): void
@@ -53,10 +78,9 @@ final class TecanProtocolTest extends TestCase
                 new TransferWithAutoWash(100, $liquidClass, $aspirateLocation, $dispenseLocation)
             );
         }
-
         self::assertSame(
             StringUtil::normalizeLineEndings(
-                'A;;;TestRackType;;barcode;100;TestLiquidClassName;;1
+                $this->initComment() . 'A;;;TestRackType;;barcode;100;TestLiquidClassName;;1
 D;;;TestRackType;;barcode1;100;TestLiquidClassName;;1
 W;
 A;;;TestRackType;;barcode;100;TestLiquidClassName;;2
@@ -101,7 +125,7 @@ W;
 
         self::assertSame(
             StringUtil::normalizeLineEndings(
-                'A;;;TestRackType;;barcode;100;TestLiquidClassName;;1
+                $this->initComment() . 'A;;;TestRackType;;barcode;100;TestLiquidClassName;;1
 D;;;TestRackType;;barcode1;100;TestLiquidClassName;;1
 W;
 A;;;TestRackType;;barcode;100;TestLiquidClassName;;2
@@ -140,7 +164,7 @@ W;
 
         self::assertSame(
             StringUtil::normalizeLineEndings(
-                'A;;;TestRackType;;barcode;100;TestLiquidClassName;;1
+                $this->initComment() . 'A;;;TestRackType;;barcode;100;TestLiquidClassName;;1
 D;;;TestRackType;;barcode1;100;TestLiquidClassName;;1
 W;
 A;;;TestRackType;;barcode;100;TestLiquidClassName;;2
@@ -257,5 +281,14 @@ R;MM;;Eppis 32x1.5 ml Cooled;5;5;DestPCR;;96 Well PCR ABI semi-skirted;1;5;24;Tr
             ),
             $tecanProtocol->buildProtocol()
         );
+    }
+
+    private function initComment(): string
+    {
+        $version = InstalledVersions::getPrettyVersion(TecanProtocol::PACKAGE_NAME);
+
+        return "C;Created by mll-lab/liquid-handling-robotics v.{$version}
+C;Date: 2022-10-05 13:34:32
+";
     }
 }